Created a separate file for helper functions, such as making text ascii printable and formatting attribute names.
Fixed issue where the player bio button was not sending the correct player's info.

Changed the window titles to be centered using 'just' and not changing the widths.

Added a images folder. Added a placeholder screen shot while trying to figure out how to get an image from a URL.